add workaround for switchbot boot loop issues on ESPHome
Users have reported boot loop issues on ESPHome with the C3 chips. This includes the one on the popular Switchbot 15A plug. 

Adding some logic to the api and to the ble tracker seems to fix this issue.

This adds the workaround to the esphome YAML for the Switchbot plug.